add layout protocol
