+++ cocos/ui/UILayout.h
@@ -26,6 +26,7 @@ THE SOFTWARE.
 #define __LAYOUT_H__
 
 #include "ui/UIWidget.h"
+#include "ui/CCLayoutProtocol.h"
 
 NS_CC_BEGIN
 
@@ -35,21 +36,12 @@ namespace ui {
  *  @js NA
  *  @lua NA
  */
-class LayoutExecutant;
-    
-class Layout : public Widget
+class Layout : public Widget, public LayoutProtocol
 {
     
     DECLARE_CLASS_GUI_INFO
     
 public:
-    enum class Type
-    {
-        ABSOLUTE,
-        VERTICAL,
-        HORIZONTAL,
-        RELATIVE
-    };
     
     enum class ClippingType
     {
@@ -303,7 +295,7 @@ protected:
     //init background image renderer.
     void addBackGroundImage();
     
-    void supplyTheLayoutParameterLackToChild(Widget* child);
+    void supplyTheLayoutParameterLackToChild(LayoutParameterProtocol* child);
     virtual Widget* createCloneInstance() override;
     virtual void copySpecialProperties(Widget* model) override;
     virtual void copyClonedWidgetChildren(Widget* model) override;
@@ -325,7 +317,6 @@ protected:
     void updateBackGroundImageColor();
     void updateBackGroundImageOpacity();
     void updateBackGroundImageRGBA();
-    LayoutExecutant* createCurrentLayoutExecutant();
     
     /**
      *get the content size of the layout, it will accumulate all its children's content size
@@ -448,7 +439,6 @@ protected:
     Vector2 _alongVector;
     GLubyte _cOpacity;
     Size _backGroundImageTextureSize;
-    Type _layoutType;
     ClippingType _clippingType;
     DrawNode* _clippingStencil;
     bool _scissorRectDirty;
@@ -477,7 +467,6 @@ protected:
     Color3B _backGroundImageColor;
     GLubyte _backGroundImageOpacity;
     
-    LayoutExecutant* _curLayoutExecutant;
     
     GLint _mask_layer_le;
     GroupCommand _groupCommand;
